fix: 파일 구조 변경 #3의 파일 구조의 컨벤션에 맞춰서 파일 구조를 수정하였습니다.
