--- lib/auth/index.ts
@@ -1,4 +1,0 @@
-import serverSideAuthentication from "@Auth/serverSideAuthentication";
-import clientSideAuthentication from "@Auth/clientSideAuthentication";
-
-export { serverSideAuthentication, clientSideAuthentication };
