--- lib/messages/constants/error/index.ts
@@ -1,4 +1,0 @@
-import NETWORK_ERROR_MESSAGE from "@Messages/constants/error/networkErrorMessage";
-import NOT_EXIST_MESSAGE from "@Messages/constants/error/notExistMessage";
-
-export { NETWORK_ERROR_MESSAGE, NOT_EXIST_MESSAGE };
