--- lib/messages/constants/error/notExistMessage.ts
@@ -1,8 +1,0 @@
-import MessageInterface from "@Messages/interfaces";
-
-const NOT_EXIST_MESSAGE: MessageInterface = {
-  ko: "에러 메세지가 존재하지 않습니다.",
-  en: "Error message does not exist.",
-};
-
-export default NOT_EXIST_MESSAGE;
