--- lib/messages/interfaces/index.ts
@@ -1,4 +1,0 @@
-import { LANGUAGE } from "@Language/language";
-
-type MessageInterface = Record<typeof LANGUAGE[keyof typeof LANGUAGE], string>;
-export default MessageInterface;
