--- lib/open-api/errors/interfaces/index.ts
@@ -1,8 +1,0 @@
-import MessageInterface from "@Messages/interfaces";
-
-type ExceptionInterface = {
-  statusCode: number;
-  requestLocation: string;
-  message: MessageInterface;
-};
-export default ExceptionInterface;
