+++ src/constants/date-fns/index.ts
@@ -1,0 +1,4 @@
+import { JUST_NOW } from "src/utility/date-fns/constants";
+
+JUST_NOW;
+export default JUST_NOW;
