+++ src/constants/messages/error/networkErrorMessage.ts
@@ -1,0 +1,7 @@
+import MessageInterface from "constants/messages/interfaces";
+
+const NETWORK_ERROR_MESSAGE: MessageInterface = {
+  ko: "서버와 연결이 끊겼습니다.",
+  en: "NetWork Error.",
+};
+export default NETWORK_ERROR_MESSAGE;
