+++ src/services/adapter/category.ts
@@ -1,0 +1,11 @@
+import * as TILog from "@til-log.lab/tilog-api";
+import axiosInstance from "services/open-api/axiosInstance";
+import { Option } from "../interface";
+
+const categoryApi = new TILog.CategoryApi(undefined, undefined, axiosInstance);
+
+export const getCategory = (options?: Option) =>
+  categoryApi.categoriesControllerGetCategories(options);
+
+export const getUserCategory = (userId: number, options?: Option) =>
+  categoryApi.categoriesControllerGetUsersCategories(userId, options);
