+++ src/services/adapter/comment.ts
@@ -1,0 +1,26 @@
+import * as TILog from "@til-log.lab/tilog-api";
+import axiosInstance from "services/open-api/axiosInstance";
+import { Option } from "../interface";
+
+const commentApi = new TILog.CommentApi(undefined, undefined, axiosInstance);
+
+export const createComment = (
+  comment: TILog.CreateCommentsRequestBodyDto,
+  options?: Option
+) => commentApi.commentsControllerCreateComment(comment, options);
+
+export const getComment = (
+  postId: string,
+  replyTo?: string,
+  options?: Option
+) => commentApi.commentsControllerGetComments(postId, replyTo, options);
+
+export const deleteComment = (
+  comment: TILog.DeleteCommentRequestDto,
+  options?: Option
+) => commentApi.commentsControllerDeleteComment(comment, options);
+
+export const updateComment = (
+  comment: TILog.UpdateCommentRequestDto,
+  options?: Option
+) => commentApi.commentsControllerUpdateComment(comment, options);
