+++ src/services/auth/index.ts
@@ -1,0 +1,4 @@
+import serverSideAuthentication from "src/services/auth/serverSideAuthentication";
+import clientSideAuthentication from "src/services/auth/clientSideAuthentication";
+
+export { serverSideAuthentication, clientSideAuthentication };
