+++ src/services/errors/interfaces/index.ts
@@ -1,0 +1,8 @@
+import MessageInterface from "constants/messages/interfaces";
+
+type ExceptionInterface = {
+  statusCode: number;
+  requestLocation: string;
+  message: MessageInterface;
+};
+export default ExceptionInterface;
